fix: altered Integer for Double in atribute 'price'

+++ src/main/java/com/lucenasoft/marketplace/models/ProductRequestDTO.java
@@ -1,4 +1,4 @@
 package com.lucenasoft.marketplace.models;
 
-public record ProductRequestDTO(String title, String imgURL, Integer price) {
+public record ProductRequestDTO(String title, String imgURL, Double price) {
 }
